PIG-1041: javac warnings: cast, fallthrough, serial
git-svn-id: https://svn.apache.org/repos/asf/hadoop/pig/trunk@829216 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/apache/pig/pen/LineageTrimmingVisitor.java
@@ -338,8 +338,7 @@ public class LineageTrimmingVisitor extends LOVisitor {
                 Set<IdentityHashSet<Tuple>> equivClassesCovered = lineageGroupToEquivClasses
                         .get(lineageGroup);
                 int numEquivClassesCovered = equivClassesCovered.size();
-                double score = ((double) numEquivClassesCovered)
-                        / ((double) weight);
+                double score = ((double) numEquivClassesCovered) / weight;
 
                 if (score > bestScore) {
